+++ src/qSort.java
@@ -1,12 +1,12 @@
 public class qSort {
-    public static void sort(float[]a , int[] b, int l, int r)
+    public static void sort(double[]a , int[] b, int l, int r)
     {
         int i = l;
         int j = r;
         if ( l >=r) return;
-        float temp ;
+        double temp ;
         int temp1 ;
-        float key = a[(l+r) /2] ;
+        double key = a[(l+r) /2] ;
         do {
             while (a[i] < key ) i++;
             while (a[j] > key ) j--;
